Make mch-fs use RepositoryAccess instead of MchRepository
Not so fun fact: I accidentally corrupted a repository while developing the changes for this commit. In MchRepository.java line 242 I accidentally provided mchRegionFilePath as the input stream, and added the DELETE_ON_CLOSE flag, which deleted the .mchrs.zst file instead of the temporary file :cry:. This effectively wiped spawn and surrounding builds from the repository, which was basically the only important thing stored in that repo. This is why I shouldn't be testing with my only copy of a real repository lol.